requirements.txt
django==5.0.0
django-sslserver==0.22
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
